style: address warnings and coding style issues
Some of these warnings are from PVS Studio.
Reported by Alexandre Duret-Lutz in #140.

* vcsn/algos/are-isomorphic.hh, vcsn/algos/dot.hh,
* vcsn/algos/reduce.hh, vcsn/algos/star-height.hh,
* vcsn/algos/to-expansion.hh, vcsn/algos/to-expression.hh:
Initialize all the members in the ctors.
Prefer the auto style.
Fix a couple of names for private members.
Remove duplicate type aliases.
Move details into detail::.
Prefer emplace.
Remove misleading const and add missing ones.

+++ vcsn/algos/dot.hh
@@ -200,10 +200,11 @@ namespace vcsn
                 bos_ << "\", shape = box";
                 if (format_ == format::latex)
                   {
-                    // Approximate a fixed width based on raw text output
+                    // Approximate a fixed width based on raw text
+                    // output.
                     std::ostringstream oss;
                     aut_->print_state_name(s, oss, format::text);
-                    size_t len = oss.str().size();
+                    float len = oss.str().size();
                     float width = 0.5f + 0.1f * (len / 2);
                     bos_ << ", fixedsize = true"
                          << ", width = " << width;
